Changing front-end style

+++ frontend/src/Pages/App/Components/Communication/Components/DeleteMessages/index.tsx
@@ -6,6 +6,7 @@ import { messagesSelectedAtom } from "../../../../../../States/messagesSelected"
 import { messagesAtom } from "../../../../../../States/messages";
 import { errorAtom } from "../../../../../../States/error";
 import styles from './DeleteMessages.module.scss';
+import { apiLoadingStatusAtom } from "States/apiLoadingStatus";
 
 
 export default function DeleteMessages() {
@@ -13,8 +14,11 @@ export default function DeleteMessages() {
     const [messagesSelectedState, setMessagesSelectedState] = useRecoilState(messagesSelectedAtom);
     const [messagesState, setMessagesState] = useRecoilState(messagesAtom);
     const errorsState = useSetRecoilState(errorAtom);
+    const [apiLoadingStatusState, setApiLoadingStatusState] = useRecoilState(apiLoadingStatusAtom);
 
     function _Error(error: any) {
+        setApiLoadingStatusState(false);
+
         if('response' in error && 'error' in error.response.data) {
             errorsState((_) => typeof error.response.data['error'] === 'string' ? [error.response.data['error']] : [...error.response.data['error']]);
         } else {
@@ -23,6 +27,9 @@ export default function DeleteMessages() {
     }
 
     function HandleClick() {
+        if(apiLoadingStatusState) return;
+        setApiLoadingStatusState(true);
+
         axios.delete(`http://127.0.0.1:8000/messages/`, {
             headers: {
                 'token': jwtToken
@@ -31,6 +38,7 @@ export default function DeleteMessages() {
                 'messages_id_to_delete': messagesSelectedState
             }
         }).then(response => {
+            setApiLoadingStatusState(false);
             setJwtToken(response.data['token']);
 
             const newMessages = messagesState.filter(message => !(messagesSelectedState.includes(message.id)));
